Updated for 1.6.2 Updated Health and Damage for 1.6.2 Fixed Horses being prevented from spawning Added spawning options for Horses in World config Fixed world configs not being generated Updated version to 5.9.9

+++ src/net/jzx7/regios/bukkit/entity/BukkitPlayer.java
@@ -69,12 +69,12 @@ public class BukkitPlayer implements RegiosPlayer {
 	}
 
 	@Override
-	public int getHealth() {
+	public double getHealth() {
 		return bukkitPlayer.getHealth();
 	}
 
 	@Override
-	public void setHealth(int h) {
+	public void setHealth(double h) {
 		bukkitPlayer.setHealth(h);
 	}
 
@@ -152,7 +152,7 @@ public class BukkitPlayer implements RegiosPlayer {
 	}
 
 	@Override
-	public void damage(int damage) {
+	public void damage(double damage) {
 		bukkitPlayer.damage(damage);
 	}
